Move `/prices`
